backend/pyproject.toml
[project]
name = "mastering-ielts"
version = "0.1.0"
description = "Mastering IELTS is a web application designed to help users practice and improve their English skills in preparation for the IELTS exam"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=5.0.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "ruff>=0.9.9",
]
